requirements.txt
streamlit
requests
pandas
numpy
matplotlib
google.generativeai
python-dotenv
json